build EVERYTHING

+++ derivations.nix
@@ -7,16 +7,24 @@ let
     rec {
       name = "Haskell";
       extension = "hs";
-      buildInputs = extraDeps:
-        assert builtins.isAttrs extraDeps;
-        let
-          inherit (pkgs) haskellPackages;
-          extras = extraDeps.hs or (const [ ]);
-          aoc = haskellPackages.callPackage ./adventofcode/default.nix { };
-          ghcPkgs = hpkgs:
-            with hpkgs;
-            [ megaparsec split MonadRandom vector aoc ] ++ extras hpkgs;
-        in [ (haskellPackages.ghcWithPackages ghcPkgs) ];
+      buildInputs = let
+        inherit (pkgs) haskellPackages;
+        aoc = haskellPackages.callPackage ./adventofcode/default.nix { };
+        ghcPkgs = hpkgs:
+          with hpkgs; [
+            MonadRandom
+            aeson
+            aoc
+            cryptonite
+            fingertree
+            lens
+            megaparsec
+            multiset
+            pipes
+            split
+            vector
+          ];
+      in [ (haskellPackages.ghcWithPackages ghcPkgs) ];
       buildPhase = ''
         ghc -O2 run.hs
       '';
@@ -24,9 +32,7 @@ let
     rec {
       name = "Assembler";
       extension = "asm";
-      buildInputs = extraDeps:
-        with pkgs;
-        [ nasm manpages gdb glibc.dev ] ++ (extraDeps.asm or [ ]);
+      buildInputs = with pkgs; [ nasm manpages gdb glibc.dev ];
       buildPhase = ''
         nasm -felf64 run.asm && ld -o run run.o
       '';
@@ -36,36 +42,46 @@ let
     with lang;
     pkgs.stdenv.mkDerivation {
       name = "aoc-${name}-${toString y}-day${toString d}";
-      buildInputs = [  ] ++ buildInputs (dayDeps y d);
       src = builtins.filterSource (path: type:
         (matches ".*.${extension}" path) || (matches ".*.txt" path))
         (dayPath y d);
-      inherit buildPhase;
+      inherit buildInputs;
+      buildPhase = if builtins.pathExists (dayLangSkipPath y d lang) then
+        "touch skip"
+      else
+        buildPhase;
       installPhase = ''
-        mkdir -p $out/bin
-        cp run $out/bin
+        mkdir -p $out
+        if ! [ -e skip ]; then
+          mkdir -p $out/bin
+          cp run $out/bin
+        fi
+        cp status $out
       '';
       doCheck = true;
       checkPhase = ''
-        touch input.txt expected.txt
-        time ./run input.txt < input.txt > result.txt
-        if diff -B -Z result.txt expected.txt; then
-          echo 'Got the expected results!'
+        if ! [ -e expected.txt ] || [ -e skip ]; then
+          echo '?' > status
         else
-          echo 'Oh-oh, tests failed!'
-          echo Got
-          cat result.txt
-          echo Expected
-          cat expected.txt
-          exit 1
+          touch input.txt
+          time ./run input.txt < input.txt > result.txt
+          if diff -B -Z result.txt expected.txt; then
+            echo 'Got the expected results!'
+            echo 'OK' > status
+          else
+            echo 'Oh-oh, tests failed!'
+            echo Got
+            cat result.txt
+            echo Expected
+            cat expected.txt
+            exit 1
+          fi
         fi
       '';
     };
   dayPath = y: d: ./. + "/${y}/day${toString d}";
-  dayDeps = y: d:
-    let p = dayPath y d + "/extra.nix";
-    in if builtins.pathExists p then import p else { };
   dayLangPath = y: d: lang: dayPath y d + "/run.${lang.extension}";
+  dayLangSkipPath = y: d: lang: dayPath y d + "/.skip.${lang.extension}";
   dayLangs = y: d:
     let hasCodeFor = lang: builtins.pathExists (dayLangPath y d lang);
     in builtins.filter hasCodeFor langs;
@@ -81,4 +97,7 @@ let
       value = dayDrvs y d;
     }) days.${y});
   }) (builtins.attrNames days));
-in drvs
+  flat = with builtins;
+    let flatten1 = xs: concatLists (map attrValues xs);
+    in flatten1 (flatten1 (attrValues drvs));
+in drvs // { all = flat; }
